offersModel & destinationsModel moved to pointsModel
