+++ src/main.js
@@ -1,9 +1,7 @@
 import { render } from './framework/render';
-import PointsApiService from './point-api-service';
+import PointsApiService from './points-api-service';
 
 import PointsModel from './model/points-model';
-import DestinationsModel from './model/destinations-model';
-import OffersModel from './model/offers-model';
 import FiltersModel from './model/filters-model';
 
 import MainPresenter from './presenter/main-presenter';
@@ -22,15 +20,6 @@ const headerElement = document.querySelector('.trip-main');
 const newPointButtonComponent = new NewPointButtonView({
   handleNewPointButtonClick: handleNewPointButtonClick,
 });
-
-const offersModel = new OffersModel(
-  { pointsApiService: new PointsApiService(END_POINT, AUTHORIZATION) }
-);
-
-const destinationsModel = new DestinationsModel(
-  { pointsApiService: new PointsApiService(END_POINT, AUTHORIZATION) }
-);
-
 const pointsModel = new PointsModel(
   { pointsApiService: new PointsApiService(END_POINT, AUTHORIZATION) }
 );
@@ -40,8 +29,6 @@ const filtersModel = new FiltersModel();
 const tripInfoPresenter = new TripInfoPresenter({
   headerContainer: headerElement,
   pointsModel,
-  offersModel,
-  destinationsModel,
 });
 
 
@@ -51,26 +38,18 @@ const filtersPresenter = new FiltersPresenter({
   pointsModel,
 });
 
-Promise
-  .all([
-    offersModel.init(),
-    destinationsModel.init()
-  ])
-  .then(() => {
-    pointsModel.init()
-      .finally(() => {
-        render(newPointButtonComponent, headerElement);
-        tripInfoPresenter.init();
-        filtersPresenter.init();
-      });
+
+pointsModel.init()
+  .finally(() => {
+    render(newPointButtonComponent, headerElement);
+    tripInfoPresenter.init();
+    filtersPresenter.init();
   });
 
 
 const mainPresenter = new MainPresenter({
   pointsContainer: pointsElement,
   pointsModel,
-  offersModel,
-  destinationsModel,
   filtersModel,
   handleNewPointDestroy,
 });
