+++ src/model/points-model.js
@@ -4,6 +4,8 @@ import { UpdateType } from '../const';
 
 export default class PointsModel extends Observable {
   #points = [];
+  #offers = [];
+  #destinations = [];
 
   #pointsApiService = null;
 
@@ -13,8 +15,16 @@ export default class PointsModel extends Observable {
     this.#pointsApiService = pointsApiService;
   }
 
+  get offers() {
+    return this.#offers;
+  }
+
   get points() {
     return this.#points;
+  }
+
+  get destinations() {
+    return this.#destinations;
   }
 
   #adaptToClient(point) {
@@ -98,9 +108,13 @@ export default class PointsModel extends Observable {
   // @------------ INIT ------------@ //
   async init() {
     try {
+      this.#offers = await this.#pointsApiService.offers;
+      this.#destinations = await this.#pointsApiService.destinations;
       const points = await this.#pointsApiService.points;
       this.#points = points.map(this.#adaptToClient);
     } catch (error) {
+      this.#offers = [];
+      this.#destinations = [];
       this.#points = [];
     }
     this._notify(UpdateType.INIT);
